+++ zero_sum_eval/analysis/calculate_ratings.py
