+++ zero_sum_eval/games/chess/chess_game.py
@@ -1,10 +1,10 @@
 import chess
 
 from zero_sum_eval.games.chess.chess_player import ChessPlayer, WHITE_KEY, BLACK_KEY
-from zero_sum_eval.player import Move
-from zero_sum_eval.game_state import GameState, InvalidMoveError, PlayerDefinition
-from zero_sum_eval.type_definitions import Action
-from zero_sum_eval.registry import GAME_REGISTRY
+from zero_sum_eval.core.player import Move
+from zero_sum_eval.core.game_state import GameState, InvalidMoveError, PlayerDefinition
+from zero_sum_eval.utils.types import Action
+from zero_sum_eval.core.registry import GAME_REGISTRY
 from typing import Dict, List
 
 @GAME_REGISTRY.register("chess")
